handle CSS vars anywhere in a property value
This enables working with CSS vars that are placed inside of other CSS functions, like rgba and hsla.  Instead of going through the `toColor` utility, the values are returned as-is.

This enables DaisyUI’s color theming to work beautifully.  I’m also amazed at how much simpler the API is compared to the TailwindCSS syntax.  Nice job, guys!